# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTerminalAccountRequest": {
        "properties": {
          "display_name": {
            "description": "ASCII terminal display name",
            "title": "Display Name",
            "type": "string"
          },
          "operator": {
            "description": "explicit operator identity",
            "title": "Operator",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "operator"
        ],
        "title": "CreateTerminalAccountRequest",
        "type": "object"
      },
...
      "Enrollment": {
        "properties": {
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "controlled_scan_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controlled Scan Time"
          },
          "controlled_scan_window_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controlled Scan Window Until"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "device_uid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Uid"
          },
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "employee_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Name"
          },
          "enrollment_id": {
            "title": "Enrollment Id",
            "type": "integer"
          },
          "fingerprint_confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint Confirmed At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "reserved_at": {
            "format": "date-time",
            "title": "Reserved At",
            "type": "string"
          },
          "reserved_by": {
            "title": "Reserved By",
            "type": "string"
          },
          "reserved_device_user_id": {
            "title": "Reserved Device User Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terminal_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Created At"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "enrollment_id",
          "employee_id",
          "device_id",
          "reserved_device_user_id",
          "status",
          "reserved_by",
          "reserved_at",
          "created_at",
          "updated_at"
        ],
        "title": "Enrollment",
        "type": "object"
      },
...
      "EnrollmentReserveResult": {
        "description": "Return of POST /api/v1/enrollments/reserve.",
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "enrollment_id": {
            "title": "Enrollment Id",
            "type": "integer"
          },
          "reserved_at": {
            "format": "date-time",
            "title": "Reserved At",
            "type": "string"
          },
          "reserved_device_user_id": {
            "title": "Reserved Device User Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "enrollment_id",
          "reserved_device_user_id",
          "status",
          "reserved_at",
          "employee_id",
          "device_id"
        ],
        "title": "EnrollmentReserveResult",
        "type": "object"
      },
      "EnrollmentTransitionResult": {
        "description": "Return of enrollment transition POSTs.",
        "properties": {
          "enrollment_id": {
            "title": "Enrollment Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "enrollment_id",
          "status"
        ],
        "title": "EnrollmentTransitionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_Enrollment_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Enrollment]",
        "type": "object"
      },
...
      "ReserveRequest": {
        "properties": {
          "device_id": {
            "description": "devices.device_id",
            "title": "Device Id",
            "type": "integer"
          },
          "employee_id": {
            "description": "Human UUID",
            "title": "Employee Id",
            "type": "string"
          },
          "operator": {
            "description": "explicit operator identity",
            "title": "Operator",
            "type": "string"
          },
          "roster_user_ids": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "optional live terminal roster IDs (read-only evidence)",
            "title": "Roster User Ids"
          }
        },
        "required": [
          "employee_id",
          "device_id",
          "operator"
        ],
        "title": "ReserveRequest",
        "type": "object"
      },
      "ScanConfirmationRequest": {
        "properties": {
          "operator": {
            "description": "explicit operator identity",
            "title": "Operator",
            "type": "string"
          },
          "scan_time": {
            "description": "controlled attendance scan time",
            "format": "date-time",
            "title": "Scan Time",
            "type": "string"
          }
        },
        "required": [
          "operator",
          "scan_time"
        ],
        "title": "ScanConfirmationRequest",
        "type": "object"
      },
...
      "TransitionRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "operator": {
            "description": "explicit operator identity",
            "title": "Operator",
            "type": "string"
          }
        },
        "required": [
          "operator"
        ],
        "title": "TransitionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "ADMS (Attendance Device Management System) UI-facing API. Backend/identity foundation is 100% complete; this layer exposes stable read contracts and gated canonical workflow wrappers for the frontend. F5 auth: operator accounts with VIEWER/OPERATOR/ADMIN roles.",
    "title": "ADMS API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/enrollments": {
      "get": {
        "operationId": "list_enrollments_api_v1_enrollments_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "employee_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Employee Id"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Enrollment_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Enrollments",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/reserve": {
      "post": {
        "operationId": "reserve_api_v1_enrollments_reserve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentReserveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reserve",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}": {
      "get": {
        "operationId": "get_enrollment_detail_api_v1_enrollments__enrollment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enrollment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Enrollment Detail",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/cancel": {
      "post": {
        "operationId": "cancel_api_v1_enrollments__enrollment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/confirm-controlled-scan": {
      "post": {
        "operationId": "confirm_scan_api_v1_enrollments__enrollment_id__confirm_controlled_scan_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Scan",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/confirm-fingerprint": {
      "post": {
        "operationId": "confirm_fingerprint_api_v1_enrollments__enrollment_id__confirm_fingerprint_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Fingerprint",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/create-terminal-account": {
      "post": {
        "operationId": "create_terminal_account_api_v1_enrollments__enrollment_id__create_terminal_account_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTerminalAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Terminal Account",
        "tags": [
          "enrollments"
        ]
      }
    },
...
    "/api/v1/enrollments/{enrollment_id}/next-actions": {
      "get": {
        "description": "Returns the valid next operator actions for an enrollment's current state.\n\nComputed from the canonical state machine (app.enrollment.ALLOWED_TRANSITIONS\n+ ENROLLMENT_ACTIONS) so the frontend never duplicates workflow logic.\nRead-only \u2014 no state mutation. Physical terminal steps (e.g. terminal\naccount creation) are intentionally absent: they happen at the terminal.",
        "operationId": "get_next_actions_api_v1_enrollments__enrollment_id__next_actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentNextActions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Next Actions",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/start-controlled-scan": {
      "post": {
        "operationId": "start_scan_window_api_v1_enrollments__enrollment_id__start_controlled_scan_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Scan Window",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{enrollment_id}/start-fingerprint-enrollment": {
      "post": {
        "operationId": "start_fingerprint_api_v1_enrollments__enrollment_id__start_fingerprint_enrollment_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "title": "Enrollment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTransitionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Fingerprint",
        "tags": [
          "enrollments"
        ]
      }
    },
...
  }
}